feat: add reusable seo component
