--- config/index.ts
@@ -1,3 +1,0 @@
-const dev = process.env.NODE_ENV !== "production"
-
-export const server = dev ? "http://localhost:3000" : "https://poached.vercel.app"
